Compatibility with symfony/process 5.0
