+++ bin/ci/phpunit-php-7-4-symfony-5-0
@@ -1,0 +1,13 @@
+#!/usr/bin/env sh
+
+set -eu
+
+readonly ROOT_DIR="$(realpath "$(dirname "$(realpath "$0")")/../..")"
+. "${ROOT_DIR}"/bin/common.inc.sh
+
+if ${IS_IN_DOCKER}; then
+    readonly PHP_BIN="php7.4"
+    readonly COMPOSER_HOME_SYMFONY="${COMPOSER_HOME_SYMFONY_5_0}"
+fi
+
+. "${ROOT_DIR}"/bin/ci/phpunit-php.common.sh
