+++ bin/ci/phpunit.php
@@ -11,9 +11,11 @@ use Symfony\Component\Console\Input\ArgvInput;
 require dirname(__DIR__, 2) . '/vendor/autoload.php';
 
 (new ParallelProcessesApplication())
+    ->addProcess(new Process([__DIR__ . '/phpunit-php-7-4-symfony-5-0']))
     ->addProcess(new Process([__DIR__ . '/phpunit-php-7-4-symfony-5-1']))
     ->addProcess(new Process([__DIR__ . '/phpunit-php-7-4-symfony-5-2']))
     ->addProcess(new Process([__DIR__ . '/phpunit-php-7-4-symfony-5-3']))
+    ->addProcess(new Process([__DIR__ . '/phpunit-php-8-0-symfony-5-0']))
     ->addProcess(new Process([__DIR__ . '/phpunit-php-8-0-symfony-5-1']))
     ->addProcess(new Process([__DIR__ . '/phpunit-php-8-0-symfony-5-2']))
     ->addProcess(new Process([__DIR__ . '/phpunit-php-8-0-symfony-5-3']))
