+++ composer.json
@@ -21,10 +21,10 @@
         }
     },
     "require": {
-        "php": "^7.4 || ^8.0",
+        "php": "^7.4||^8.0",
         "steevanb/php-typed-array": "^3.2",
         "symfony/console": "^5.2",
-        "symfony/process": "^5.1"
+        "symfony/process": "^5.0"
     },
     "require-dev": {
         "ext-simplexml": "*",
